refactor(core): change module relations
