+++ modules/@ergonode/attributes/src/defaults/attributes/index.js
@@ -2,6 +2,14 @@
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE for license details.
  */
+export const SKU_MODEL_ID = 'SKU';
+
+export const SKU_MODEL = {
+    id: SKU_MODEL_ID,
+    key: SKU_MODEL_ID,
+    value: SKU_MODEL_ID,
+    hint: SKU_MODEL_ID,
+};
 
 export const TYPES = {
     TEXT: 'TEXT',
