+++ modules/@ergonode/comments/src/config/imports.js
@@ -1,0 +1,11 @@
+/*
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE for license details.
+ */
+export const Tabs = {
+    ProductCommentsTab: () => import('@Comments/components/Tabs/ProductCommentsTab').then(m => m.default || m),
+};
+
+export const Icons = {
+    Product: () => import('@Core/components/Icons/Menu/IconProduct'),
+};
