--- modules/@ergonode/product-templates/src/defaults/product/index.js
@@ -1,12 +1,0 @@
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE for license details.
- */
-export const SKU_MODEL_ID = 'SKU';
-
-export const SKU_MODEL = {
-    id: SKU_MODEL_ID,
-    key: SKU_MODEL_ID,
-    value: SKU_MODEL_ID,
-    hint: SKU_MODEL_ID,
-};
